First version of GitHub Status indicator
This commit introduces the code for a GitHub Status informer app that will
send a message to a Slack channel (via incoming Webhook) when GitHub Status
reports an incident.
There will be another message sent to the channel when the status changes or
the incident gets resolved.

This way, developers at dxw will know in real time when GitHub is down.

It is designed to run regularly in the background and only send messages on
Slack when there is a change on GitHub status.