S3 bucket and logic added
The app didn't work properly once it was on Heroku due to the yaml file that
stores the information about github status being re-started from default
every time Heroku kicked off the scheduler on a one-off dyno.

This commit solves that problem by storing the yaml file on an AWS S3 bucket
that Heroku will be able to access to load the content and update the status
every time it runs on the scheduler.